vet_rag_project/requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
torch
jq
tiktoken
openai
python-dotenv
gradio
kss
pyyaml
scikit-learn
transformers
ragas
langchain-openai
datasets
huggingface_hub
